refactor(bot/autoplay): extract candidateCollector↔spotifyRecommender helpers to candidateContracts.ts
Break Cycle B by extracting the function-level mutual dependencies
(shouldIncludeCandidate, upsertScoredCandidate) from candidateCollector
and moving them to a new neutral module candidateContracts.ts.

Both modules now import these shared helpers from candidateContracts
instead of each other, severing the back-edge:
  candidateCollector → spotifyRecommender → candidateContracts

- Created packages/bot/src/utils/music/autoplay/candidateContracts.ts
- Updated spotifyRecommender.ts to import from candidateContracts
- candidateCollector.ts re-exports from candidateContracts for backward compat
